// packages/graph/src/types/node.types.ts
export enum ENodeType {
  BINDING_SINGLE = "bindingSingle",
  BINDING_MULTI = "bindingMulti",
  INDICATOR = "indicator",
  OPERATOR = "operator",
  TASK = "task",
  GROUP = "group",
  BRANCH_PLACEHOLDER = "branchPlaceholder",
  BINDING_PLACEHOLDER = "bindingPlaceholder",
}

export enum EEdgeType {
  EDGE_WITH_BUTTON = "edgeWithButton",
}